requirements.txt
django-crispy-forms==1.9.0
gunicorn==19.10.0
